rename resources package to nitric
