+++ nitric/options.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package resources
+package nitric
 
 import "github.com/nitrictech/go-sdk/faas"
